# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/products/{id}": {
      "get": {
        "operationId": "ProductsController_findOne",
        "summary": "Get product by ID",
        "description": "Retrieve a specific product by ID with all details including variations",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "includeVariations",
            "required": false,
            "in": "query",
            "description": "Include product variations in response",
            "schema": {
              "default": true,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "patch": {
        "operationId": "ProductsController_update",
        "summary": "Update product",
        "description": "Update product information",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "ProductsController_remove",
        "summary": "Delete product",
        "description": "Delete a product",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully"
          },
          "404": {
            "description": "Product not found"
          },
          "409": {
            "description": "Cannot delete product with existing decants"
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
    "/product-variations": {
      "post": {
        "operationId": "ProductVariationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductVariationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "product-variations"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "get": {
        "operationId": "ProductVariationsController_findAll",
        "summary": "Get variants",
        "description": "Get all variants or filter by product ID",
        "parameters": [
          {
            "name": "productId",
            "required": false,
            "in": "query",
            "description": "Filter variants by product ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of variants",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductVariationResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "product-variations"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/product-variations/{id}": {
      "get": {
        "operationId": "ProductVariationsController_findOne",
        "summary": "Get variant by ID",
        "description": "Retrieve detailed information about a specific product variation/variant including option values",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product Variation ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Variant found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductVariationResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Variant not found"
          }
        },
        "tags": [
          "product-variations"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "patch": {
        "operationId": "ProductVariationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductVariationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "product-variations"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "ProductVariationsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "product-variations"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "E-commerce API",
    "description": "E-commerce API with NestJS - Products, Categories, Users, and Authentication",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "ProductResponseDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateProductVariationDto": {
        "type": "object",
        "properties": {}
      },
      "ProductVariationResponseDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateProductVariationDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}